Switch mouse pos to i16.
This is because the mouse position can be reported outside the game
canvas in the case of a drag gesture.

+++ runtimes/web/src/runtime.js
@@ -32,8 +32,8 @@ export class Runtime {
     }
 
     setMouse (x, y, buttons) {
-        this.data.setUint8(constants.ADDR_MOUSE_X, x);
-        this.data.setUint8(constants.ADDR_MOUSE_Y, y);
+        this.data.setInt16(constants.ADDR_MOUSE_X, x);
+        this.data.setInt16(constants.ADDR_MOUSE_Y, y);
         this.data.setUint8(constants.ADDR_MOUSE_BUTTONS, buttons);
     }
